# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "metering-svc-hnc-metering.apps.blackwhale.cloud.hancom.com",
    "basePath": "/api",
    "paths": {
        "/count/{container}": {
            "get": {
                "description": "특정기간동안 검색된 namespace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "namespace count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "container name",
                        "name": "container",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "end",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        },
        "/cpu/{container}": {
            "get": {
                "description": "현재 cpu 사용량",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "cpu usage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "container name",
                        "name": "container",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        }
    }
}